feat: add roles management page
